+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-02-01
+
+### Added
+
+- `MaskEnableFlags` bitflags struct
+- `set_critical_alert_latch(bool)` mutator method
+- `set_warning_alert_latch(bool)` mutator method
+- `get_channels_enabled(&[bool])` method
+- `set_channels_enabled(&mut [bool])` mutator method
+
+### Changed
+
+- `read_alert_flags()` method now returns `MaskEnableFlags` bitflags struct
+- `read_alert_flags()` method now allows preserving the alert flags
+- Documentation
+
 ## [0.2.1] - 2023-02-01
 
 ### Fixed